GobStopper (For YouTube) - First working implementation
This is the first working implementation of the addon. It currently
works for YouTube only, and can only pause all the open tabs. This
works across all windows, though a Firefox window has to be focussed
for the F4 hotkey to work).

The "Ambox" icon is currently used as the icon/logo for this addon.
A nicer one can be put in place later. For now, this rough POC will
do for the motivating use case for this addon.