Stop every video element on each page visited
Previously, only the first video element encountered would be muted.
On YouTube there is only a single video element so this will be fine,
but on other sites, this may not be the case. Loosen this restriction
so that this addon will work on sites with multiple video elements
(where the first one encountered may not actually be the one playing
sounds).